+++ bot/types/Author.py
