+++ bot/types/RGB.py
