+++ main.py
@@ -1,4 +1,4 @@
-from bot import main
+from bot.bot import main
 
 if __name__ == "__main__":
     main()
